Add BUILD.bazel file
