+++ maven_install.json
@@ -1,7 +1,7 @@
 {
   "__AUTOGENERATED_FILE_DO_NOT_MODIFY_THIS_FILE_MANUALLY": "THERE_IS_NO_DATA_ONLY_ZUUL",
-  "__INPUT_ARTIFACTS_HASH": 1425304572,
-  "__RESOLVED_ARTIFACTS_HASH": -1806191256,
+  "__INPUT_ARTIFACTS_HASH": 355745105,
+  "__RESOLVED_ARTIFACTS_HASH": 308276413,
   "artifacts": {
     "args4j:args4j": {
       "shasums": {
@@ -1095,6 +1095,13 @@
         "sources": "3fa1a3e5031e20fd7ab34d5ae9dff576836c99d9c07b55f66faf6ae9a43e3bc8"
       },
       "version": "5.6"
+    },
+    "com.rabbitmq:amqp-client": {
+      "shasums": {
+        "jar": "f5df000445bbab12f650d4cd9224c57c2e416eb75a7b80f5734c670fbf932b1e",
+        "sources": "de7f700b6922dba8fdd84f84d43531b061085521329686009b37818ec13e4bd6"
+      },
+      "version": "5.22.0"
     },
     "com.squareup.okhttp3:logging-interceptor": {
       "shasums": {
@@ -4840,6 +4847,9 @@
       "org.apache.commons:commons-lang3",
       "org.apache.commons:commons-text"
     ],
+    "com.rabbitmq:amqp-client": [
+      "org.slf4j:slf4j-api"
+    ],
     "com.squareup.okhttp3:logging-interceptor": [
       "com.squareup.okhttp3:okhttp",
       "org.jetbrains.kotlin:kotlin-stdlib-jdk8"
@@ -7387,6 +7397,18 @@
       "com.opencsv.processor",
       "com.opencsv.stream.reader",
       "com.opencsv.validators"
+    ],
+    "com.rabbitmq:amqp-client": [
+      "com.rabbitmq.client",
+      "com.rabbitmq.client.impl",
+      "com.rabbitmq.client.impl.nio",
+      "com.rabbitmq.client.impl.recovery",
+      "com.rabbitmq.client.observation",
+      "com.rabbitmq.client.observation.micrometer",
+      "com.rabbitmq.tools",
+      "com.rabbitmq.tools.json",
+      "com.rabbitmq.tools.jsonrpc",
+      "com.rabbitmq.utility"
     ],
     "com.squareup.okhttp3:logging-interceptor": [
       "okhttp3.logging"
@@ -12360,6 +12382,8 @@
       "com.ongres.stringprep:stringprep:jar:sources",
       "com.opencsv:opencsv",
       "com.opencsv:opencsv:jar:sources",
+      "com.rabbitmq:amqp-client",
+      "com.rabbitmq:amqp-client:jar:sources",
       "com.squareup.okhttp3:logging-interceptor",
       "com.squareup.okhttp3:logging-interceptor:jar:sources",
       "com.squareup.okhttp3:okhttp",
