+++ src/main/kotlin/org/wfanet/measurement/securecomputation/controlplane/client/v1alpha/BUILD.bazel
@@ -1,0 +1,14 @@
+load("@wfa_rules_kotlin_jvm//kotlin:defs.bzl", "kt_jvm_library")
+
+package(default_visibility = [
+    "//src/main/kotlin/org/wfanet/measurement/securecomputation:__subpackages__",
+    "//src/test/kotlin/org/wfanet/measurement/securecomputation:__subpackages__",
+])
+
+kt_jvm_library(
+    name = "rabbit_mq_handler",
+    srcs = ["RabbitMQHandler.kt"],
+    deps = [
+        "@maven//:com_rabbitmq_amqp_client"
+    ],
+)
